requirements.txt
# pip3 install -r requirements.txt
sentence-transformers
cassandra-driver 
matplotlib 
sentence-transformers